tunnel: fix: Only attempt reconnects if previous session was closed
I found this bug when my internet connection went away and `usque` had to reconnect. If a client sends a large packet before the connection could be fully set up and tunnel MTU properly negotiated, we could receive a non fatal DATAGRAM too large error. Previously this was treated fatal, but that's wrong.
We are only closing now if the connect-ip connection is closed.